requirements.txt
flask
openai
ortools
anthropic
google-cloud-aiplatform
pyyaml
streamlit
requests
python-dotenv
xlsxwriter
